# [CWS] improvements to path reducer