### What does this PR do?

This PR implements multiple optimizations to the path reducer:
- split the different pattern regexps
- add hint and pre-check to skip regexp all together if useless
- using non capturing groups
- splitting some regexp for better hint and literal checks

#### Benchmark:

before:
```
goos: linux
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree
BenchmarkPathsReducer_ReducePath/proc_1-4         	  600784	      1859 ns/op	     553 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_2-4         	  779632	      1380 ns/op	     545 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_3-4         	  606025	      1950 ns/op	      48 B/op	       1 allocs/op
BenchmarkPathsReducer_ReducePath/proc_4-4         	  587793	      1823 ns/op	     552 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_task_1-4    	  517554	      2151 ns/op	     833 B/op	       6 allocs/op
BenchmarkPathsReducer_ReducePath/proc_task_1#01-4 	  357202	      3201 ns/op	     560 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/cgroup_1-4       	  128049	      9181 ns/op	    1009 B/op	       6 allocs/op
BenchmarkPathsReducer_ReducePath/proc_cgroup_1-4  	  111108	     10632 ns/op	    1409 B/op	       8 allocs/op
BenchmarkPathsReducer_ReducePath/container_id_1-4 	  188516	      6149 ns/op	     576 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/block_device_1-4 	  350930	      3185 ns/op	     592 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/block_device_2-4 	 1589284	       757.8 ns/op	     562 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/k8s_service_account_secret_1-4         	  460522	      2423 ns/op	     593 B/op	       4 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree	14.974s
```

after:
```
goos: linux
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree
BenchmarkPathsReducer_ReducePath/proc_1-4         	 3240608	       367.4 ns/op	     362 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_2-4         	 3328880	       355.3 ns/op	     354 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_3-4         	 7948915	       148.9 ns/op	      64 B/op	       1 allocs/op
BenchmarkPathsReducer_ReducePath/proc_4-4         	 3503229	       341.8 ns/op	     362 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/proc_task_1-4    	 1768755	       679.3 ns/op	     675 B/op	       7 allocs/op
BenchmarkPathsReducer_ReducePath/proc_task_1#01-4 	 2308273	       513.0 ns/op	     370 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/cgroup_1-4       	  275413	      4090 ns/op	     834 B/op	       7 allocs/op
BenchmarkPathsReducer_ReducePath/proc_cgroup_1-4  	  235053	      4936 ns/op	    1283 B/op	      10 allocs/op
BenchmarkPathsReducer_ReducePath/container_id_1-4 	  612571	      1822 ns/op	     385 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/block_device_1-4 	 1579119	       759.7 ns/op	     401 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/block_device_2-4 	 2178434	       547.9 ns/op	     369 B/op	       4 allocs/op
BenchmarkPathsReducer_ReducePath/k8s_service_account_secret_1-4         	  353930	      3267 ns/op	     401 B/op	       4 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree	18.116s
```
### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.